requirements.txt
absl-py==2.1.0
amqp==5.2.0
asgiref==3.8.1
astunparse==1.6.3
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.1
cron-descriptor==1.4.3
cycler==0.12.1
Django==5.0.6
django-celery-beat==2.6.0
django-rest-framework==0.1.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
flatbuffers==24.3.25
fonttools==4.53.0
gast==0.5.4
geographiclib==2.0
geopy==2.4.1
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.7
keras==3.3.3
kiwisolver==1.4.5
kombu==5.3.7
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
prompt_toolkit==3.0.46
protobuf==4.25.3
psycopg2-binary==2.9.9
Pygments==2.18.0
pygrib==2.1.5
pyparsing==3.1.2
pyproj==3.6.1
python-crontab==3.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.6.0
redis==5.0.5
requests==2.32.3
rich==13.7.1
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
    